+++ utils/schedule_func.py
@@ -1,0 +1,28 @@
+# -*- coding:utf-8 -*-
+# @Time     : 2019-01-12 21:57
+# @Author   : 胡远
+# @Github   : https://github.com/QuixoteHY
+# @Email    : 1290482442@qq.com
+# @Describe : 函数调度工具
+
+from quixote import loop
+
+
+class CallLaterOnce(object):
+    def __init__(self, func, *a, **kw):
+        self._func = func
+        self._a = a
+        self._kw = kw
+        self._call = None
+
+    def schedule(self, delay=0):
+        if self._call is None:
+            self._call = loop.call_later(delay, self)
+
+    def cancel(self):
+        if self._call:
+            self._call.cancel()
+
+    def __call__(self):
+        self._call = None
+        return self._func(*self._a, **self._kw)
